requirements.txt
fastapi
uvicorn
pytest
coverage
codecov
requests
pydantic
fastapi-sqlalchemy
sqlalchemy
psycopg2-binary
alembic
gunicorn
PyYAML
